tests/requirements.txt
pytest
pytest-xdist
pyyaml
boto3
